fix: add class app assets for organization and add folder shared for the bottom navigation bar to remove from main
