+++ lib/core/app_assets.dart
@@ -1,11 +1,13 @@
-const iconLitecoin = 'assets/icons/litecoin_icon.png';
-const iconBitcoin = 'assets/icons/bitcoin_vector.png';
-const iconEthereum = 'assets/icons/ethereum_icon.png';
+class AppAssets {
+  final iconLitecoin = 'assets/icons/litecoin_icon.png';
+  final iconBitcoin = 'assets/icons/bitcoin_vector.png';
+  final iconEthereum = 'assets/icons/ethereum_icon.png';
 
-const iconWarrenOutlined = 'assets/icons/warren_outlined_icon.png';
-const iconWarren = 'assets/icons/warren_icon.png';
+  final iconWarrenOutlined = 'assets/icons/warren_outlined_icon.png';
+  final iconWarren = 'assets/icons/warren_icon.png';
 
-const iconCryptoOutlined = 'assets/icons/crypto_outlined_icon.png';
-const iconCrypto = 'assets/icons/crypto_icon.png';
+  final iconCryptoOutlined = 'assets/icons/crypto_outlined_icon.png';
+  final iconCrypto = 'assets/icons/crypto_icon.png';
 
-const imageWorkInProgress = 'assets/images/work_in_progress.png';
+  final imageWorkInProgress = 'assets/images/work_in_progress.png';
+}
